+++ Frends.HTTP.Request/Frends.HTTP.Request/Definitions/Header.cs
@@ -8,10 +8,12 @@ public class Header
     /// <summary>
     /// Name of header.
     /// </summary>
+    /// <example>Content-Type</example>
     public string Name { get; set; }
 
     /// <summary>
     /// Value of header.
     /// </summary>
+    /// <example>application/json</example>
     public string Value { get; set; }
 }