+++ Frends.HTTP.Request/Frends.HTTP.Request/Definitions/Options.cs
@@ -188,4 +188,19 @@ public class Options
     /// <example>Default</example>
     [DefaultValue(SslVersion.Default)]
     public SslVersion SslProtocolVersion { get; set; } = SslVersion.Default;
+
+    /// <summary>
+    /// Whether to throw an error on failure.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Overrides the error message on failure.
+    /// </summary>
+    /// <example>HTTP request failed: connection refused</example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 }